refactor: consolidate duplicate UE5 project into single clean structure
Merge two parallel UE5 project structures (root Source/ and nested
ShardsOfDawn/Source/) into one unified project at root level using
standard UE5 Public/Private layout.

Changes:
- Move GAS abilities, combat, camera, co-op, UI systems from nested
  ShardsOfDawn/ into root Source/ShardsOfDawn/Public+Private/
- Remove duplicate nested ShardsOfDawn/ directory entirely
- Update .uproject with Editor module, OnlineSubsystem, CommonUI plugins
- Update Build.cs with networking, AI, UI dependencies
- Enhance DefaultEngine.ini (Lumen, Nanite, collision channels, networking)
- Add DefaultGameplayTags.ini (all ability, combat, chapter tags)
- Add root .gitignore for UE5 artifacts
- Update README with consolidated structure and quick-start guide
- Fix include paths for relocated puzzle actor files
