+++ Config/DefaultEngine.ini
@@ -2,7 +2,43 @@
 GameDefaultMap=/Game/Maps/PrototypeRoom
 EditorStartupMap=/Game/Maps/PrototypeRoom
 GlobalDefaultGameMode=/Game/Game/GameModes/BP_SodGameMode.BP_SodGameMode_C
+GameInstanceClass=/Script/ShardsOfDawn.SODGameInstance
 
 [/Script/Engine.Engine]
 +ActiveGameNameRedirects=(OldGameName="TP_Blank",NewGameName="/Script/ShardsOfDawn")
 +ActiveGameNameRedirects=(OldGameName="/Script/TP_Blank",NewGameName="/Script/ShardsOfDawn")
+
+[/Script/Engine.GameEngine]
++NetDriverDefinitions=(DefName="GameNetDriver",DriverClassName="OnlineSubsystemUtils.IpNetDriver",DriverClassNameFallback="OnlineSubsystemUtils.IpNetDriver")
+
+[/Script/Engine.RendererSettings]
+r.DefaultFeature.AutoExposure=True
+r.Lumen.DiffuseIndirectLighting.HardwareRayTracing=True
+r.RayTracing=True
+r.Nanite=True
+
+[/Script/HardwareTargeting.HardwareTargetingSettings]
+TargetedHardwareClass=Desktop
+AppliedTargetedHardwareClass=Desktop
+DefaultGraphicsPerformance=Maximum
+AppliedDefaultGraphicsPerformance=Maximum
+
+[/Script/Engine.CollisionProfile]
++Profiles=(Name="LightInteract",CollisionEnabled=QueryOnly,ObjectTypeName="GameTraceChannel1",CustomResponses=((Channel="WorldStatic",Response=ECR_Overlap),(Channel="WorldDynamic",Response=ECR_Overlap)))
++Profiles=(Name="ShadowInteract",CollisionEnabled=QueryOnly,ObjectTypeName="GameTraceChannel2",CustomResponses=((Channel="WorldStatic",Response=ECR_Overlap),(Channel="WorldDynamic",Response=ECR_Overlap)))
+
++DefaultChannelResponses=(Channel=ECC_GameTraceChannel1,DefaultResponse=ECR_Ignore,bTraceType=True,bStaticObject=False,Name="LightInteract")
++DefaultChannelResponses=(Channel=ECC_GameTraceChannel2,DefaultResponse=ECR_Ignore,bTraceType=True,bStaticObject=False,Name="ShadowInteract")
+
+[/Script/OnlineSubsystemUtils.IpNetDriver]
+NetServerMaxTickRate=60
+
+[/Script/Engine.PhysicsSettings]
+DefaultGravityZ=-980.0
+DefaultTerminalVelocity=4000.0
+
+[URL]
+GameName=ShardsOfDawn
+
+[/Script/WindowsTargetPlatform.WindowsTargetSettings]
+DefaultGraphicsRHI=DirectX 12
