--- ShardsOfDawn/Config/DefaultGame.ini
@@ -1,31 +1,0 @@
-[/Script/EngineSettings.GeneralProjectSettings]
-ProjectID=ShardsOfDawn2024
-ProjectName=Shards of Dawn
-CompanyName=ShardsOfDawn Studio
-CompanyDistinguishedName=com.shardsofdawn
-Description=A premium co-op action-adventure game featuring Light/Shadow dual-state mechanics. Two players must cooperate to solve puzzles, defeat enemies, and rebuild a shattered world.
-Homepage=
-SupportContact=
-CopyrightNotice=Copyright Shards of Dawn. All Rights Reserved.
-bShouldWindowPreserveAspectRatio=True
-bUseBorderlessWindow=False
-bStartVR=False
-
-[/Script/UnrealEd.ProjectPackagingSettings]
-+DirectoriesToAlwaysStageAsUFS=(Path="Puzzles")
-+DirectoriesToAlwaysStageAsUFS=(Path="Abilities")
-BuildConfiguration=PPBC_Shipping
-StagingDirectory=(Path="")
-FullRebuild=False
-ForDistribution=False
-IncludeDebugFiles=False
-BlueprintNativizationMethod=Disabled
-bShareMaterialShaderCode=True
-bCompressed=True
-
-[Internationalization]
-+CulturesToStage=en
-+CulturesToStage=vi
-+CulturesToStage=ja
-+CulturesToStage=ko
-+CulturesToStage=zh-Hans
