--- ShardsOfDawn/Config/DefaultInput.ini
@@ -1,28 +1,0 @@
-[/Script/Engine.InputSettings]
--AxisConfig=(AxisKeyName="Gamepad_LeftX",DeadZone=0.25,Sensitivity=1.0,Exponent=1.0,bInvert=False)
-+AxisConfig=(AxisKeyName="Gamepad_LeftX",DeadZone=0.15,Sensitivity=1.0,Exponent=1.0,bInvert=False)
--AxisConfig=(AxisKeyName="Gamepad_LeftY",DeadZone=0.25,Sensitivity=1.0,Exponent=1.0,bInvert=False)
-+AxisConfig=(AxisKeyName="Gamepad_LeftY",DeadZone=0.15,Sensitivity=1.0,Exponent=1.0,bInvert=False)
--AxisConfig=(AxisKeyName="Gamepad_RightX",DeadZone=0.25,Sensitivity=1.0,Exponent=1.0,bInvert=False)
-+AxisConfig=(AxisKeyName="Gamepad_RightX",DeadZone=0.15,Sensitivity=1.0,Exponent=1.0,bInvert=False)
--AxisConfig=(AxisKeyName="Gamepad_RightY",DeadZone=0.25,Sensitivity=1.0,Exponent=1.0,bInvert=False)
-+AxisConfig=(AxisKeyName="Gamepad_RightY",DeadZone=0.15,Sensitivity=1.0,Exponent=1.0,bInvert=False)
-bUseMouseForTouch=False
-DefaultViewportMouseCaptureMode=CapturePermanently_IncludingInitialMouseDown
-bEnableMouseSmoothing=False
-DefaultViewportMouseLockMode=LockOnCapture
-
-[/Script/EnhancedInput.EnhancedInputDeveloperSettings]
-PlatformSettings=(("Windows", "/Game/Input/PlatformInput_PC"))
-
-; NOTE: Actual Input Actions and Mapping Contexts are defined as UE5 assets in Content/Input/
-; The following are action name references for documentation:
-;
-; IA_Move - 2D Axis (WASD / Left Stick)
-; IA_Look - 2D Axis (Mouse / Right Stick)
-; IA_Jump - Digital (Space / A Button)
-; IA_Interact - Digital (E / X Button)
-; IA_AbilityPrimary - Digital (LMB / Right Trigger)
-; IA_AbilitySecondary - Digital (RMB / Left Trigger)
-; IA_Ping - Digital (Middle Mouse / D-Pad Up)
-; IA_Pause - Digital (Escape / Start)
