--- ShardsOfDawn/ShardsOfDawn.uproject
@@ -1,66 +1,0 @@
-{
-	"FileVersion": 3,
-	"EngineAssociation": "5.4",
-	"Category": "",
-	"Description": "Shards of Dawn - A premium co-op action-adventure game featuring Light/Shadow dual-state mechanics.",
-	"Modules": [
-		{
-			"Name": "ShardsOfDawn",
-			"Type": "Runtime",
-			"LoadingPhase": "Default",
-			"AdditionalDependencies": [
-				"Engine",
-				"GameplayAbilities",
-				"EnhancedInput"
-			]
-		},
-		{
-			"Name": "ShardsOfDawnEditor",
-			"Type": "Editor",
-			"LoadingPhase": "PostEngineInit"
-		}
-	],
-	"Plugins": [
-		{
-			"Name": "GameplayAbilities",
-			"Enabled": true
-		},
-		{
-			"Name": "EnhancedInput",
-			"Enabled": true
-		},
-		{
-			"Name": "Niagara",
-			"Enabled": true
-		},
-		{
-			"Name": "OnlineSubsystem",
-			"Enabled": true
-		},
-		{
-			"Name": "OnlineSubsystemUtils",
-			"Enabled": true
-		},
-		{
-			"Name": "OnlineSubsystemNull",
-			"Enabled": true
-		},
-		{
-			"Name": "CommonUI",
-			"Enabled": true
-		},
-		{
-			"Name": "ControlFlows",
-			"Enabled": true
-		},
-		{
-			"Name": "ModelViewViewModel",
-			"Enabled": true
-		}
-	],
-	"TargetPlatforms": [
-		"Windows",
-		"PS5",
-		"XSX"
-	]
-}
