--- ShardsOfDawn/Source/ShardsOfDawn/Core/ShardsOfDawnModule.cpp
@@ -1,17 +1,0 @@
-// Copyright Shards of Dawn. All Rights Reserved.
-
-#include "Core/ShardsOfDawnModule.h"
-
-#define LOCTEXT_NAMESPACE "FShardsOfDawnModule"
-
-void FShardsOfDawnModule::StartupModule()
-{
-}
-
-void FShardsOfDawnModule::ShutdownModule()
-{
-}
-
-#undef LOCTEXT_NAMESPACE
-
-IMPLEMENT_PRIMARY_GAME_MODULE(FShardsOfDawnModule, ShardsOfDawn, "ShardsOfDawn");
