--- ShardsOfDawn/Source/ShardsOfDawn/Core/ShardsOfDawnModule.h
@@ -1,13 +1,0 @@
-// Copyright Shards of Dawn. All Rights Reserved.
-
-#pragma once
-
-#include "CoreMinimal.h"
-#include "Modules/ModuleManager.h"
-
-class FShardsOfDawnModule : public IModuleInterface
-{
-public:
-	virtual void StartupModule() override;
-	virtual void ShutdownModule() override;
-};
