--- ShardsOfDawn/Source/ShardsOfDawnEditor.Target.cs
@@ -1,15 +1,0 @@
-// Copyright Shards of Dawn. All Rights Reserved.
-
-using UnrealBuildTool;
-using System.Collections.Generic;
-
-public class ShardsOfDawnEditorTarget : TargetRules
-{
-	public ShardsOfDawnEditorTarget(TargetInfo Target) : base(Target)
-	{
-		Type = TargetType.Editor;
-		DefaultBuildSettings = BuildSettingsVersion.V4;
-		IncludeOrderVersion = EngineIncludeOrderVersion.Unreal5_4;
-		ExtraModuleNames.AddRange(new string[] { "ShardsOfDawn" });
-	}
-}
