+++ Source/ShardsOfDawn/Public/Abilities/SODGameplayAbility.h
